d/aws_ses_active_receipt_rule_set: Move tests to their own source file.
