+++ internal/service/ses/active_receipt_rule_set_data_source_test.go
@@ -1,0 +1,51 @@
+package ses_test
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/aws/aws-sdk-go/service/ses"
+	sdkacctest "github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-provider-aws/internal/acctest"
+)
+
+func testAccActiveReceiptRuleSetDataSource_basic(t *testing.T) {
+	rName := sdkacctest.RandomWithPrefix(acctest.ResourcePrefix)
+	resourceName := "data.aws_ses_active_receipt_rule_set.test"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			acctest.PreCheck(t)
+			testAccPreCheck(t)
+			testAccPreCheckSESReceiptRule(t)
+		},
+		ErrorCheck:   acctest.ErrorCheck(t, ses.EndpointsID),
+		Providers:    acctest.Providers,
+		CheckDestroy: testAccCheckSESActiveReceiptRuleSetDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccActiveReceiptRuleSetDataSourceConfig(rName),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckActiveReceiptRuleSetExists(resourceName),
+					acctest.CheckResourceAttrRegionalARN(resourceName, "arn", "ses", fmt.Sprintf("receipt-rule-set/%s", rName)),
+				),
+			},
+		},
+	})
+}
+
+func testAccActiveReceiptRuleSetDataSourceConfig(name string) string {
+	return fmt.Sprintf(`
+resource "aws_ses_receipt_rule_set" "test" {
+  rule_set_name = %[1]q
+}
+
+resource "aws_ses_active_receipt_rule_set" "test" {
+  rule_set_name = aws_ses_receipt_rule_set.test.rule_set_name
+}
+data "aws_ses_active_receipt_rule_set" "test" {
+  depends_on = [aws_ses_active_receipt_rule_set.test]
+}
+`, name)
+}
